Split out coordinate transforms
